address comment + fix tests
